[refactor] 인프라 호출 과정 리팩터링 (#1045)
* refactor: 플레이어 강퇴 로직 PlayerService로 이동 및 메서드 분리

- PlayerKickConsumer가 RoomEventService 대신 PlayerService를 사용하도록 수정
- PlayerService 내 강퇴 이벤트 로직 분리(publishKickPlayerEvent, kickPlayer)
- RoomQueryService에 existsPlayer 메서드 추가
- 관련 테스트 코드 및 컨트롤러 호출 로직 수정

* refactor: 플레이어 목록 업데이트 로직 PlayerService로 이동 및 메서드 분리

- PlayerListUpdateConsumer가 RoomEventService 대신 PlayerService를 사용하도록 수정
- PlayerService 내 목록 업데이트 로직 분리(updatePlayers)
- RoomEventService에서 목록 업데이트 관련 메서드 제거
- 관련 호출부 로직 및 로그 메시지 수정

* refactor: PlayerReady 로직 RoomEventService에서 RoomService로 통합

- PlayerReadyEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- PlayerReadyConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 및 중복 코드 제거

* refactor: 미니게임 목록 업데이트 로직 RoomService로 통합

- MiniGameSelectEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- MiniGameSelectConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 및 중복 코드 제거

* refactor: QR 코드 상태 처리 로직 RoomEventService에서 RoomService로 이동

- QrCodeStatusEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- QrCodeStatusConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 제거 및 코드 정리

* refactor: 방 생성 로직 RoomEventService에서 RoomService로 이동

- RoomCreateEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RoomCreateConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 및 코드 정리

* refactor: 방 입장 로직 RoomEventService에서 RoomService로 이동

- RoomJoinEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RoomJoinConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 제거 및 코드 정리

* refactor: 룰렛 표시 로직 RoomEventService에서 RoomService로 이동

- RouletteShowEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RouletteShowConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 제거 및 코드 정리

* refactor: 룰렛 스핀 로직 RoomEventService에서 RoomService로 이동

- RouletteSpinEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RouletteSpinConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- DataInitializable 인터페이스 제거
- 불필요한 import 제거 및 코드 정리

* refactor: 플레이어 강퇴 이벤트 처리 방식 수정

- PlayerService의 publishKickPlayerEvent 메서드 제거 및 kickPlayer 메서드로 통합
- PlayerKickEvent 객체 생성하도록 로직 변경
- 관련 테스트 코드 수정 및 정리

* refactor: 변수 선언에 final 추가 및 불필요한 메서드 제거

- 주요 클래스와 메서드에서 변경되지 않는 변수에 final 선언 추가
- 사용되지 않는 forceCleanup 메서드 제거
- 코드 가독성과 유지보수성을 위한 소소한 리팩토링 진행

* refactor: 메서드 명 변경 및 관련 코드 수정

- PlayerService의 publishKickPlayerEvent 메서드를 checkAndKickPlayer로 변경
- 메서드 호출부 및 관련 테스트 코드 수정
- 로그 메시지와 메서드 호출 로직 업데이트
