+++ backend/src/main/java/coffeeshout/room/application/service/PlayerService.java
@@ -4,6 +4,7 @@ import coffeeshout.global.redis.stream.StreamKey;
 import coffeeshout.global.redis.stream.StreamPublisher;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.event.PlayerKickEvent;
+import coffeeshout.room.domain.event.PlayerListUpdateEvent;
 import coffeeshout.room.domain.player.Player;
 import coffeeshout.room.domain.player.PlayerName;
 import coffeeshout.room.domain.service.RoomCommandService;
@@ -11,6 +12,7 @@ import coffeeshout.room.domain.service.RoomQueryService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 @Slf4j
@@ -20,12 +22,13 @@ public class PlayerService {
 
     private final RoomQueryService roomQueryService;
     private final RoomCommandService roomCommandService;
+    private final ApplicationEventPublisher eventPublisher;
     private final StreamPublisher streamPublisher;
 
-    public boolean kickPlayer(String joinCode, String playerName) {
+    public boolean checkAndKickPlayer(String joinCode, String playerName) {
         log.info("JoinCode[{}] 플레이어 강퇴 명령 처리 - 플레이어: {}", joinCode, playerName);
 
-        boolean exists = roomCommandService.removePlayer(new JoinCode(joinCode), new PlayerName(playerName));
+        final boolean exists = roomQueryService.existsPlayer(new JoinCode(joinCode), new PlayerName(playerName));
 
         if (exists) {
             final PlayerKickEvent event = new PlayerKickEvent(joinCode, playerName);
@@ -35,7 +38,21 @@ public class PlayerService {
         return exists;
     }
 
+    public void kickPlayer(PlayerKickEvent event) {
+        log.info("JoinCode[{}] 플레이어 강퇴 이벤트 처리 - 플레이어: {}", event.joinCode(), event.playerName());
+
+        roomCommandService.removePlayer(new JoinCode(event.joinCode()), new PlayerName(event.playerName()));
+
+        eventPublisher.publishEvent(new PlayerListUpdateEvent(event.joinCode()));
+    }
+
     public List<Player> getPlayers(String joinCode) {
         return roomQueryService.getPlayers(new JoinCode(joinCode));
+    }
+
+    public void updatePlayers(PlayerListUpdateEvent event) {
+        log.info("JoinCode[{}] 플레이어 목록 업데이트 이벤트 처리", event.joinCode());
+
+        eventPublisher.publishEvent(event);
     }
 }
