--- backend/src/main/java/coffeeshout/room/application/service/RoomEventService.java
@@ -1,169 +1,0 @@
-package coffeeshout.room.application.service;
-
-import coffeeshout.room.domain.JoinCode;
-import coffeeshout.room.domain.Room;
-import coffeeshout.room.domain.RoomState;
-import coffeeshout.room.domain.event.MiniGameSelectEvent;
-import coffeeshout.room.domain.event.PlayerKickEvent;
-import coffeeshout.room.domain.event.PlayerListUpdateEvent;
-import coffeeshout.room.domain.event.PlayerReadyEvent;
-import coffeeshout.room.domain.event.QrCodeStatusEvent;
-import coffeeshout.room.domain.event.RoomCreateEvent;
-import coffeeshout.room.domain.event.RoomJoinEvent;
-import coffeeshout.room.domain.event.RouletteShowEvent;
-import coffeeshout.room.domain.event.RouletteShownEvent;
-import coffeeshout.room.domain.event.RouletteSpinEvent;
-import coffeeshout.room.domain.event.RouletteWinnerEvent;
-import coffeeshout.room.domain.player.PlayerName;
-import coffeeshout.room.domain.player.Winner;
-import coffeeshout.room.domain.service.RoomCommandService;
-import coffeeshout.room.infra.messaging.RoomEventWaitManager;
-import coffeeshout.room.infra.persistence.RoulettePersistenceService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class RoomEventService {
-
-    private final RoomCommandService roomCommandService;
-    private final DelayedRoomRemovalService delayedRoomRemovalService;
-    private final ApplicationEventPublisher eventPublisher;
-    private final RoomEventWaitManager roomEventWaitManager;
-    private final RoulettePersistenceService roulettePersistenceService;
-    private final RouletteService rouletteService;
-
-    /**
-     * 미니게임 목록을 업데이트하고, WebSocket 브로드캐스트를 위해 이벤트를 재발행합니다.
-     */
-    public void updateMiniGames(MiniGameSelectEvent event) {
-        log.info("JoinCode[{}] 미니게임 목록 업데이트 이벤트 처리 - 호스트: {}, 미니게임 종류: {}",
-                event.joinCode(),
-                event.hostName(),
-                event.miniGameTypes()
-        );
-
-        roomCommandService.updateMiniGames(
-                new JoinCode(event.joinCode()),
-                new PlayerName(event.hostName()),
-                event.miniGameTypes()
-        );
-
-        eventPublisher.publishEvent(event);
-    }
-
-    public void kickPlayer(PlayerKickEvent event) {
-        log.info("JoinCode[{}] 플레이어 강퇴 이벤트 처리 - 플레이어: {}", event.joinCode(), event.playerName());
-        JoinCode joinCode = new JoinCode(event.joinCode());
-
-        roomCommandService.removePlayer(joinCode, new PlayerName(event.playerName()));
-
-        eventPublisher.publishEvent(new PlayerListUpdateEvent(event.joinCode()));
-    }
-
-    public void updatePlayers(PlayerListUpdateEvent event) {
-        log.info("JoinCode[{}] 플레이어 목록 업데이트 이벤트 처리", event.joinCode());
-
-        eventPublisher.publishEvent(event);
-    }
-
-    public void readyPlayer(PlayerReadyEvent event) {
-        log.info("JoinCode[{}] 플레이어 준비 상태 변경 이벤트 처리 - 플레이어: {}, 준비 상태: {}",
-                event.joinCode(),
-                event.playerName(),
-                event.isReady()
-        );
-
-        roomCommandService.readyPlayer(new JoinCode(event.joinCode()),
-                new PlayerName(event.playerName()),
-                event.isReady()
-        );
-
-        eventPublisher.publishEvent(new PlayerListUpdateEvent(event.joinCode()));
-    }
-
-    public void joinRoom(RoomJoinEvent event) {
-        log.info("JoinCode[{}] 게스트 방 입장 이벤트 처리 - 게스트 이름: {}",
-                event.joinCode(),
-                event.guestName()
-        );
-
-        try {
-            final Room room = roomCommandService.joinGuest(
-                    new JoinCode(event.joinCode()),
-                    new PlayerName(event.guestName())
-            );
-
-            roomEventWaitManager.notifySuccess(event.eventId(), room);
-        } catch (Exception e) {
-            roomEventWaitManager.notifyFailure(event.eventId(), e);
-            throw e;
-        }
-    }
-
-    public void createRoom(RoomCreateEvent event) {
-        log.info("JoinCode[{}] 방 생성 이벤트 처리 - 호스트 이름: {}",
-                event.joinCode(),
-                event.hostName()
-        );
-
-        roomCommandService.saveIfAbsentRoom(
-                new JoinCode(event.joinCode()),
-                new PlayerName(event.hostName())
-        );
-
-        delayedRoomRemovalService.scheduleRemoveRoom(new JoinCode(event.joinCode()));
-    }
-
-    public void showRoulette(RouletteShowEvent event) {
-        log.info("JoinCode[{}] 룰렛 화면 표시 이벤트 처리", event.joinCode());
-
-        final RoomState roomState = rouletteService.showRoulette(event.joinCode());
-
-        roulettePersistenceService.saveRoomStatus(event);
-
-        eventPublisher.publishEvent(new RouletteShownEvent(event.joinCode(), roomState));
-    }
-
-    public void handleQrCodeStatus(QrCodeStatusEvent event) {
-        log.info(
-                "QR 코드 완료 이벤트 수신: eventId={}, joinCode={}, status={}",
-                event.eventId(), event.joinCode(), event.status()
-        );
-
-        switch (event.status()) {
-            case SUCCESS -> {
-                log.info(
-                        "QR 코드 완료 이벤트 처리 완료 (SUCCESS): eventId={}, joinCode={}, url={}",
-                        event.eventId(), event.joinCode(), event.qrCodeUrl()
-                );
-                roomCommandService.assignQrCode(new JoinCode(event.joinCode()), event.qrCodeUrl());
-                eventPublisher.publishEvent(event);
-            }
-            case ERROR -> {
-                log.info(
-                        "QR 코드 완료 이벤트 처리 완료 (ERROR): eventId={}, joinCode={}",
-                        event.eventId(), event.joinCode()
-                );
-                roomCommandService.assignQrCodeError(new JoinCode(event.joinCode()));
-                eventPublisher.publishEvent(event);
-            }
-            default -> log.error(
-                    "처리할 수 없는 QR 코드 상태: eventId={}, joinCode={}, status={}",
-                    event.eventId(), event.joinCode(), event.status()
-            );
-        }
-    }
-
-    public void spinRoulette(RouletteSpinEvent event) {
-        log.info("JoinCode[{}] 룰렛 스핀 이벤트 처리 - 당첨자: {}", event.joinCode(), event.winner().name().value());
-
-        final Winner winner = event.winner();
-        roulettePersistenceService.saveRouletteResult(event);
-
-        eventPublisher.publishEvent(new RouletteWinnerEvent(event.joinCode(), winner));
-    }
-}
