+++ backend/src/main/java/coffeeshout/room/infra/messaging/consumer/MiniGameSelectConsumer.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.infra.messaging.consumer;
 
-import coffeeshout.room.application.service.RoomEventService;
+import coffeeshout.room.application.service.RoomService;
 import coffeeshout.room.domain.event.MiniGameSelectEvent;
 import java.util.function.Consumer;
 import lombok.RequiredArgsConstructor;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class MiniGameSelectConsumer implements Consumer<MiniGameSelectEvent> {
 
-    private final RoomEventService roomEventService;
+    private final RoomService roomService;
 
     @Override
     public void accept(MiniGameSelectEvent event) {
-        roomEventService.updateMiniGames(event);
+        roomService.updateMiniGames(event);
     }
 }
