+++ backend/src/main/java/coffeeshout/room/infra/messaging/consumer/PlayerReadyConsumer.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.infra.messaging.consumer;
 
-import coffeeshout.room.application.service.RoomEventService;
+import coffeeshout.room.application.service.RoomService;
 import coffeeshout.room.domain.event.PlayerReadyEvent;
 import java.util.function.Consumer;
 import lombok.RequiredArgsConstructor;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class PlayerReadyConsumer implements Consumer<PlayerReadyEvent> {
 
-    private final RoomEventService roomEventService;
+    private final RoomService roomService;
 
     @Override
     public void accept(PlayerReadyEvent event) {
-        roomEventService.readyPlayer(event);
+        roomService.readyPlayer(event);
     }
 }
