+++ backend/src/main/java/coffeeshout/room/ui/RoomRestController.java
@@ -110,7 +110,7 @@ public class RoomRestController implements RoomApi {
             @PathVariable String joinCode,
             @PathVariable String playerName
     ) {
-        final boolean exists = playerService.kickPlayer(joinCode, playerName);
+        final boolean exists = playerService.checkAndKickPlayer(joinCode, playerName);
 
         if (exists) {
             return ResponseEntity.noContent().build();
