+++ backend/src/test/java/coffeeshout/room/application/service/PlayerServiceTest.java
@@ -9,6 +9,7 @@ import coffeeshout.global.exception.custom.NotExistElementException;
 import coffeeshout.global.redis.stream.StreamPublisher;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
+import coffeeshout.room.domain.event.PlayerKickEvent;
 import coffeeshout.room.domain.player.Player;
 import coffeeshout.room.domain.player.PlayerName;
 import coffeeshout.room.domain.service.RoomCommandService;
@@ -47,7 +48,8 @@ class PlayerServiceTest extends ServiceTest {
             JoinCode joinCode = createdRoom.getJoinCode();
 
             // when
-            playerService.kickPlayer(joinCode.getValue(), hostName);
+            PlayerKickEvent event = new PlayerKickEvent(joinCode.getValue(), hostName);
+            playerService.kickPlayer(event);
 
             // then
             assertThat(roomService.roomExists(joinCode.getValue())).isFalse();
@@ -63,7 +65,7 @@ class PlayerServiceTest extends ServiceTest {
             joinGuest(joinCode, "게스트1");
 
             // when
-            playerService.kickPlayer(joinCode.getValue(), hostName);
+            playerService.checkAndKickPlayer(joinCode.getValue(), hostName);
 
             // then
             assertThat(roomService.roomExists(joinCode.getValue())).isTrue();
@@ -82,7 +84,8 @@ class PlayerServiceTest extends ServiceTest {
                 joinGuest(joinCode, guestName);
 
                 // when
-                playerService.kickPlayer(joinCode.getValue(), guestName);
+                PlayerKickEvent event = new PlayerKickEvent(joinCode.getValue(), guestName);
+                playerService.kickPlayer(event);
 
                 // then
                 assertThat(roomService.roomExists(joinCode.getValue())).isTrue();
@@ -99,7 +102,7 @@ class PlayerServiceTest extends ServiceTest {
                 JoinCode joinCode = createdRoom.getJoinCode();
 
                 // when & then
-                assertThatThrownBy(() -> playerService.kickPlayer(joinCode.getValue(), "존재하지_않는_게스트"))
+                assertThatThrownBy(() -> playerService.checkAndKickPlayer(joinCode.getValue(), "존재하지_않는_게스트"))
                         .isInstanceOf(InvalidArgumentException.class);
             }
 
@@ -111,7 +114,7 @@ class PlayerServiceTest extends ServiceTest {
                 JoinCode joinCode = createdRoom.getJoinCode();
 
                 // when & then
-                assertThatThrownBy(() -> playerService.kickPlayer(joinCode.getValue(), null))
+                assertThatThrownBy(() -> playerService.checkAndKickPlayer(joinCode.getValue(), null))
                         .isInstanceOf(InvalidArgumentException.class);
             }
         }
@@ -119,14 +122,15 @@ class PlayerServiceTest extends ServiceTest {
         @Test
         void 존재하지_않는_방_코드로_제거_시도_시_예외가_발생한다() {
             // when & then
-            assertThatThrownBy(() -> playerService.kickPlayer("ABCD", "플레이어"))
+            PlayerKickEvent event = new PlayerKickEvent("ABCD", "플레이어");
+            assertThatThrownBy(() -> playerService.kickPlayer(event))
                     .isInstanceOf(NotExistElementException.class);
         }
 
         @Test
         void null_방_코드로_제거_시도_시_예외가_발생한다() {
             // when & then
-            assertThatThrownBy(() -> playerService.kickPlayer(null, "플레이어"))
+            assertThatThrownBy(() -> playerService.checkAndKickPlayer(null, "플레이어"))
                     .isInstanceOf(InvalidArgumentException.class);
         }
     }
@@ -159,7 +163,8 @@ class PlayerServiceTest extends ServiceTest {
             JoinCode joinCode = createdRoom.getJoinCode();
 
             // 모든 플레이어 제거 (방도 제거됨)
-            playerService.kickPlayer(joinCode.getValue(), hostName);
+            PlayerKickEvent event = new PlayerKickEvent(joinCode.getValue(), hostName);
+            playerService.kickPlayer(event);
 
             // when & then
             assertThatThrownBy(() -> playerService.getPlayers(joinCode.getValue()))
